Give more chance for goalie to move while intercepting the ball
In the previous code the goalie was hardly moving even after the a kick
was predicted. This was becuase the goalie was taking action only when
the ball has moved 0.2m.
Now moving him every 0.1m, and also keep moving in the same direction
for 0.5 seconds to make sure the ball has settled

Signed-off-by: Adham Abozaeid <adham.abozaeid@gmail.com>